+++ InvertBinaryTree/InvertBinaryTree/Program.cs
@@ -1,21 +1,8 @@
-﻿using System.Collections.Generic;
-
+﻿
 namespace InvertBinaryTree
 {
     public class Program
     {
-        internal class BinaryTree
-        {
-            internal int value;
-            internal BinaryTree leftChild;
-            internal BinaryTree rightChild;
-
-            internal BinaryTree(int value)
-            {
-                this.value = value;
-            }
-        }
-
         public static void Main(string[] args)
         {
             BinaryTree tree = new BinaryTree(1);
@@ -25,34 +12,8 @@ namespace InvertBinaryTree
             tree.leftChild.rightChild = new BinaryTree(5);
             tree.rightChild.leftChild = new BinaryTree(6);
 
-            Program p = new Program();
-            p.InvertBTree(tree);
-        }
-
-        internal void InvertBTree(BinaryTree tree)
-        {
-            List<BinaryTree> queue = new List<BinaryTree>();
-            queue.Add(tree);
-            int index = 0;
-            while (index < queue.Count)
-            {
-                BinaryTree current = queue[index];
-
-                if (current.leftChild != null)
-                    queue.Add(current.leftChild);
-                if (current.rightChild != null)
-                    queue.Add(current.rightChild);
-
-                index += 1;
-                SwapLeftAndRight(current);
-            }
-        }
-
-        internal void SwapLeftAndRight(BinaryTree tree)
-        {
-            BinaryTree left = tree.leftChild;
-            tree.leftChild = tree.rightChild;
-            tree.rightChild = left;
+            Invert invert = new Invert();
+            invert.InvertBTree(tree);
         }
     }
 }
